feat: get user by id controller
