+++ api/controllers/userController.js
@@ -1,19 +1,72 @@
-import { User } from '../models/index.js';
+import * as bcrypt from 'bcrypt';
+import jwt from 'jwt-simple';
+import UserModel from '../models/User.js';
+import config from '../config/index.js';
 
-/**
+const create = async (req, res) => {
+  try {
+    const hashedPassword = await bcrypt.hash(req.body.password, 10);
+    const newUser = {
+      ...req.body,
+      password: hashedPassword,
+    };
+    const createdUser = await UserModel.create(newUser);
+    res.status(201).json(createdUser);
+  } catch (err) {
+    return res.status(500).json({ err });
+  }
+};
 
-  Acceder a mi información (buscar user por id sacado del token)
+const login = async (req, res) => {
+  try {
+    const user = await UserModel.findOne(req.email);
+    if (!user) {
+      return res.status(401).json({
+        msg: 'Credenciales erróneas',
+      });
+    }
+    const compared = await bcrypt.compare(req.password);
+    if (!compared) {
+      return res.status(401).json({
+        msg: 'Bad credentials',
+      });
+    }
+    user.password = undefined;
+    const token = jwt.encode(user, config.token.secret);
+    return res.json({
+      msg: 'Login satisfactorio',
+      token,
+    });
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({
+      msg: 'Error al hacer login',
+    });
+  }
+};
 
-  Si eres admin puedes pasar un id para buscar al usuario
-
-  NUNCA mostrar contraseña
-
- */
+const updateById = async (req, res) => {
+  if (req.user.role != 'admin') {
+    return res.status(402).json({
+      msg: 'No auth'
+    })
+  }
+  try {
+    const { id } = req.params;
+    const user = await UserModel.findByIdAndUpdate(id, req.body);
+    return res.json({
+      msg: 'User actualizado',
+      user,
+    });
+  } catch (error) {
+    return returnError('Error al actualizar user');
+  }
+};
 
 const getUserByIdToken = async (req, res) => {
   const { id } = req.params;
   try {
-    const user = await User.findById(id);
+    const user = await UserModel.findById(id);
 
     if (req.user.role === 'admin') {
       return res.json({
@@ -24,8 +77,9 @@ const getUserByIdToken = async (req, res) => {
 
     if (req.user.role != 'admin') {
       return res.json({
-        msg: 'Permisos no válidos'
-      });
+        msg: 'Usuario encontrado',
+        user: req.user
+      }) 
     }
   } catch (error) {
     return res.status(500).json({
@@ -35,4 +89,4 @@ const getUserByIdToken = async (req, res) => {
   }
 };
 
-export { getUserByIdToken };
+export { create, login, updateById, getUserByIdToken };
